requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
requests>=2.31
python-dotenv>=1.0
watchdog>=4.0
werkzeug>=3.0
